Refactor: simplify project structure
