+++ src/eShopApplicationOnWeb.Application/Contracts/Infrastructure/Identity/Dtos/Login/LoginRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Identity.Dtos.Auth.Login
+namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Identity.Dtos.Login
 {
     public class LoginRequest
     {
