+++ src/eShopApplicationOnWeb.Application/Contracts/Infrastructure/Identity/Dtos/Login/LoginResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Identity.Dtos.Auth.Login
+﻿namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Identity.Dtos.Login
 {
     public class LoginResponse
     {
