+++ src/eShopApplicationOnWeb.Application/Contracts/Infrastructure/Identity/Dtos/Register/RegisterResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Identity.Dtos.Auth.Register
+﻿namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Identity.Dtos.Register
 {
     public class RegisterResponse
     {
