--- src/eShopApplicationOnWeb.Application/Contracts/Infrastructure/Services/CurrentUser/ICurrentUserService.cs
@@ -1,7 +1,0 @@
-﻿namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Services.CurrentUser
-{
-    public interface ICurrentUserService
-    {
-        string? GetUserId();
-    }
-}
