+++ src/eShopApplicationOnWeb.Application/Contracts/Infrastructure/Services/ICurrentUserService.cs
@@ -1,0 +1,7 @@
+﻿namespace eShopApplicationOnWeb.Application.Contracts.Infrastructure.Services
+{
+    public interface ICurrentUserService
+    {
+        string? GetUserId();
+    }
+}
