+++ src/eShopApplicationOnWeb.Infrastructure/Extensions/InfrastructureServiceExtensions.cs
@@ -1,4 +1,6 @@
-﻿using eShopApplicationOnWeb.Infrastructure.Identity.Extensions;
+﻿using eShopApplicationOnWeb.Application.Contracts.Infrastructure.Services;
+using eShopApplicationOnWeb.Infrastructure.Identity.Extensions;
+using eShopApplicationOnWeb.Infrastructure.Services;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -9,6 +11,8 @@ namespace eShopApplicationOnWeb.Infrastructure.Extensions
         public static IServiceCollection AddInfrastructureServices(this IServiceCollection services, IConfiguration configuration)
         {
             services.AddIdentityServices(configuration);
+
+            services.AddScoped<ICurrentUserService, CurrentUserService>();
 
             return services;
         }
