+++ src/eShopApplicationOnWeb.Infrastructure/Identity/Security/Jwt/JwtTokenGenerator.cs
@@ -1,12 +1,11 @@
 ﻿using eShopApplicationOnWeb.Infrastructure.Identity.Models;
-using eShopApplicationOnWeb.Infrastructure.Identity.Security.Settings;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 
-namespace eShopApplicationOnWeb.Infrastructure.Identity.Security.Tokens.Jwt
+namespace eShopApplicationOnWeb.Infrastructure.Identity.Security.Jwt
 {
     public class JwtTokenGenerator : IJwtTokenGenerator
     {
